+++ xl_auth/oauth/__init__.py
@@ -3,4 +3,4 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from . import views  # noqa
+from . import client, grant, token, views  # noqa
