+++ xl_auth/oauth/client/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""The OAuth Client package."""
+"""The OAuth Client sub-package."""
 
 from __future__ import absolute_import, division, print_function, unicode_literals
