+++ xl_auth/oauth/grant/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""The OAuth Grant module."""
+"""The OAuth Grant sub-package."""
 
 from __future__ import absolute_import, division, print_function, unicode_literals
