+++ xl_auth/oauth/token/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""The token module."""
+"""The OAuth Token sub-package."""
 
 from __future__ import absolute_import, division, print_function, unicode_literals
